requirements.txt
absl-py==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpha_vantage==3.0.0
altgraph==0.17.4
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
awscli==1.42.52
awsebcli==3.25.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
blessed==1.22.0
blis==0.7.9
boto3==1.40.52
botocore==1.40.52
Brotli==1.1.0
cachetools==5.2.0
catalogue==2.0.8
cement==2.10.14
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
contourpy==1.1.1
cryptography==46.0.2
cssselect2==0.8.0
customtkinter==5.1.2
cycler==0.12.1
cymem==2.0.7
dacite==1.9.2
darkdetect==0.8.0
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.7
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.3
django-countries==7.6.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
docutils==0.19
easyocr==1.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fabric==3.2.2
faiss-cpu==1.13.2
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.12.4
Flask==2.2.2
flatbuffers==22.12.6
fonttools==4.43.1
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2023.10.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.7.0
hmmlearn==0.3.3
html5lib==1.1
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.5.1
invoke==2.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.3.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langchain-core==1.2.16
langchain-text-splitters==1.1.1
langcodes==3.3.0
langsmith==0.7.9
lazy_loader==0.4
libclang==14.0.6
lightgbm==4.6.0
lxml==5.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.0
MouseInfo==0.1.3
moviepy==1.0.3
mplcursors==0.6
mpmath==1.3.0
msal==1.34.0
multidict==6.1.0
multitasking==0.0.11
murmurhash==1.0.9
narwhals==1.24.1
networkx==3.2
newsapi-python==0.2.7
ninja==1.13.0
nltk==3.8.1
numerai-tools==0.2.3
numerapi==2.19.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.99.9
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openpyxl==3.1.4
opt-einsum==3.3.0
orjson==3.11.7
packaging==24.2
pandas==1.5.3
pandas-datareader==0.10.0
pandas-stubs==2.2.2.240603
paramiko==4.0.0
pathspec==0.12.1
pathy==0.10.1
pdfminer.six==20260107
peewee==3.17.5
pefile==2023.2.7
pillow==11.2.1
pipenv==2023.10.24
platformdirs==3.11.0
plotly==6.0.0
pocketsphinx==5.0.1
pokemontcgsdk==3.4.0
preshed==3.0.8
proglog==0.1.10
propcache==0.2.1
protobuf==3.19.6
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pyclipper==1.3.0.post6
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyJWT==2.10.1
PyMsgBox==1.0.9
PyNaCl==1.6.0
pyparsing==3.1.1
pyperclip==1.8.2
pyphen==0.17.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pyspark==4.0.1
python-bidi==0.6.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pythonnet==3.0.5
pytweening==1.0.7
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.7.2
s3transfer==0.14.0
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==5.2.3
shapely==2.1.2
Shimmy==2.0.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
sqlparse==0.5.3
srsly==2.4.6
stable_baselines3==2.5.0
starlette==0.46.2
stripe==12.4.0
sympy==1.13.1
ta-lib==0.6.1
tenacity==9.1.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.5.0
tflearn==0.5.0
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2025.3.30
tinycss2==1.4.0
tinyhtml5==2.0.0
tkinterdnd2==0.3.0
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.64.1
transformers==4.48.1
typer==0.7.0
types-pytz==2024.1.0.20240417
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ucimlrepo==0.0.7
urllib3==1.26.13
uuid_utils==0.14.1
uvicorn==0.34.2
virtualenv==20.24.6
wasabi==1.1.1
watchfiles==1.0.5
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.2
whitenoise==6.11.0
wrapt==1.14.1
xgboost==2.1.3
xxhash==3.6.0
yarl==1.18.3
yfinance==0.2.38
zopfli==0.2.3.post1
zstandard==0.25.0
